Split plugin utilities into component utilities and general app utilities for hot-reloader happiness
