+++ src/extend/pluginPreprocessing.js
@@ -1,27 +1,4 @@
-import update from 'lodash/update';
-import { connect } from 'react-redux';
 import { validatePlugin } from './pluginValidation';
-import CompanionWindowRegistry from '../lib/CompanionWindowRegistry';
-
-/**
- * Returns a mapping from targets to plugins and modes
- *
- * @param {Array} plugins
- * @return {Object} - looks like:
- *
- *  {
- *    'WorkspacePanel': {
- *      wrap:     [plugin3, ...],
- *      add:      [plugin4, ...],
- *    },
- *    ...
- *  }
- */
-export function createTargetToPluginMapping(plugins) {
-  return plugins.reduce((map, plugin) => (
-    update(map, [plugin.target, plugin.mode], x => [...x || [], plugin])
-  ), {});
-}
 
 /** */
 export function filterValidPlugins(plugins) {
@@ -31,25 +8,9 @@ export function filterValidPlugins(plugins) {
 }
 
 /** */
-export function connectPluginsToStore(plugins) {
-  return plugins.map(plugin => (
-    { ...plugin, component: connectPluginComponent(plugin) }
-  ));
-}
-
-/** */
 export function addPluginReducersToStore(store, createRootReducer, plugins) {
   const pluginReducers = getReducersFromPlugins(plugins);
   store.replaceReducer(createRootReducer(pluginReducers));
-}
-
-/** */
-export function addPluginsToCompanionWindowsRegistry(plugins) {
-  plugins.filter(p => p.companionWindowKey).forEach((plugin) => {
-    CompanionWindowRegistry[plugin.companionWindowKey] = plugin.component;
-  });
-
-  return CompanionWindowRegistry;
 }
 
 /** */
@@ -68,11 +29,6 @@ function logInvalidPlugins(plugins) {
   plugins.forEach(plugin => (
     console.log(`Mirador: Plugin ${plugin.name} is not valid and was rejected.`)
   ));
-}
-
-/** Connect plugin component to state */
-function connectPluginComponent(plugin) {
-  return connect(plugin.mapStateToProps, plugin.mapDispatchToProps)(plugin.component);
 }
 
 /**  */
